Refactor process attributes and umask handling for portability
Unified platform-specific process attributes and umask logic into dedicated files for better maintainability across Unix and Windows. Introduced `setSocketUmask` and adjusted its behavior to align with platform-specific capabilities. Removed redundant code and improved clarity in the daemon setup process.
