+++ internal/daemon/windows_compat.go
@@ -12,3 +12,9 @@ import (
 func setPlatformProcAttr(cmd *exec.Cmd) {
 	cmd.SysProcAttr = &syscall.SysProcAttr{}
 }
+
+// setSocketUmask is a no-op on Windows as umask is not supported
+func setSocketUmask(d *Daemon) func() {
+	d.logger.Debug("Umask operations not applicable on Windows")
+	return func() {}
+}
